Started separating view models
