+++ KnittingAssistant/View/MainWindow.xaml.cs
@@ -18,7 +18,6 @@ namespace KnittingAssistant.View
         private void MainWindow_Loaded(object sender, RoutedEventArgs e)
         {
             (DataContext as MainViewModel).ImageArea = imageArea;
-            (DataContext as MainViewModel).PropertyArea = propertyArea;
             (DataContext as MainViewModel).ToolbarArea = toolbarArea;
         }
     }
