+++ KnittingAssistant/View/userControls/PropertyArea.xaml.cs
@@ -1,4 +1,6 @@
-﻿using System.Windows.Controls;
+﻿using KnittingAssistant.ViewModel;
+using System.Windows;
+using System.Windows.Controls;
 
 namespace KnittingAssistant.View.userControls
 {
@@ -7,9 +9,19 @@ namespace KnittingAssistant.View.userControls
     /// </summary>
     public partial class PropertyArea : UserControl
     {
+        private PropertyAreaViewModel propertyAreaViewModel;
+
         public PropertyArea()
         {
             InitializeComponent();
+            this.Loaded += PropertyArea_Loaded;
+
+            propertyAreaViewModel = new PropertyAreaViewModel();
+        }
+
+        private void PropertyArea_Loaded(object sender, RoutedEventArgs e)
+        {
+            this.DataContext = propertyAreaViewModel;
         }
     }
 }
